platform: tegra: mc: Fix Cert-C error handling bug
In function arb_intr casting time from unsigned long long to unsigned
int without checking its value may result in lost or misinterpreted
data. Fix this by changing data type of variable.

CID 387213
Bug 3512545

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2740618
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/platform/tegra/mc/mcerr-t21.c
@@ -377,7 +377,7 @@ static void mcerr_default_print(const struct mc_error *err,
 static void arb_intr(void)
 {
 	u64 time;
-	u32 time_diff_ms;
+	u64 time_diff_ms;
 	unsigned long flags;
 
 	spin_lock_irqsave(&arb_intr_info.lock, flags);
